# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  coa.AccountTax:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      isTaxable:
        type: boolean
      name:
        type: string
      rate:
        type: number
      updatedAt:
        type: string
    type: object
  coa.AccountType:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  coa.RqCreateChartOfAccountOfAccount:
    properties:
      account_tax_id:
        minimum: 1
        type: integer
      account_type_id:
        minimum: 1
        type: integer
      code:
        maximum: 9999
        minimum: 100
        type: integer
      is_system:
        type: boolean
      name:
        maxLength: 255
        type: string
    required:
    - account_tax_id
    - account_type_id
    - code
    - name
    type: object
  coa.RqUpdateCharOfAccountOfAccount:
    properties:
      account_tax_id:
        minimum: 1
        type: integer
      account_type_id:
        minimum: 1
        type: integer
      code:
        maximum: 9999
        minimum: 100
        type: integer
      name:
        maxLength: 255
        type: string
    type: object
  coa.RsChartOfAccount:
    properties:
      account_tax_id:
        type: integer
      account_type_id:
        type: integer
      code:
        type: integer
      created_at:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      practitioner_id:
        type: string
      updated_at:
        type: string
    type: object
...
  response.RsError:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
  util.RsList:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Backend API for acareca
  title: Backend API
  version: 1.0.0
paths:
...
  /coa/account-taxes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: List all account tax types
      tags:
      - coa
  /coa/account-taxes/{id}:
    get:
      parameters:
      - description: Account Tax ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coa.AccountTax'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get account tax by ID
      tags:
      - coa
  /coa/account-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: List all account types
      tags:
      - coa
  /coa/account-types/{id}:
    get:
      parameters:
      - description: Account Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coa.AccountType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get account type by ID
      tags:
      - coa
  /coa/chart-of-account:
    get:
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by code
        in: query
        name: code
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Sort field
        in: query
        name: sort_by
        type: string
      - description: Order direction (ASC/DESC)
        in: query
        name: order_by
        type: string
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: List chart of accounts for practitioner
      tags:
      - coa
    post:
      consumes:
      - application/json
      parameters:
      - description: COA Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/coa.RqCreateChartOfAccountOfAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/coa.RsChartOfAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Create a new chart of account
      tags:
      - coa
  /coa/chart-of-account/{id}:
    delete:
      parameters:
      - description: Chart of Account UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Delete chart of account
      tags:
      - coa
    get:
      parameters:
      - description: Chart of Account UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coa.RsChartOfAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get chart of account by ID
      tags:
      - coa
    put:
      consumes:
      - application/json
      parameters:
      - description: Chart of Account UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated COA Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/coa.RqUpdateCharOfAccountOfAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coa.RsChartOfAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Update an existing chart of account
      tags:
      - coa
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    description: Type "Bearer <your_token>" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"